refactor public visual report
