+++ application/services/visual_report/transform_responses/transform_visual_sheets_to_html_with_case.rb
